requirements.txt
tweepy
kafka-python
flask
